Create Displaying North East South West for 1 time step with leaflet.ipynb
This creates a colour-map of the temperature/rainfull specififed for an area sepcified by the north, east, south and west values at a given point in time. The use of ipyleaflet enables the incorporation of the interactive map and using GEO_json to get the geographical features on the map.