Create Colour map comparison of 2 time steps.ipynb
This creates 2 colour-maps of either the temperature/rainfull specififed for an area sepcified by the north, east, south and west values at a given point in time given to different time steps. The use of ipyleaflet enables the incorporation of the interactive map and using GEO_json to get the geographical features on the map.